+++ gazebo/gui/GLWidget.cc
@@ -758,6 +758,7 @@ void GLWidget::OnMouseReleaseNormal()
           ((modelHighlighted && !rightButton) || linkHighlighted))
       {
         selectVis = linkVis;
+        this->selectionLevel = SelectionLevels::LINK;
       }
       // Select model
       else
@@ -767,6 +768,7 @@ void GLWidget::OnMouseReleaseNormal()
           this->DeselectAllVisuals();
 
         selectVis = modelVis;
+        this->selectionLevel = SelectionLevels::MODEL;
       }
       this->SetSelectedVisual(selectVis);
       event::Events::setSelectedEntity(selectVis->GetName(), "normal");
@@ -781,7 +783,8 @@ void GLWidget::OnMouseReleaseNormal()
         }
         else if (selectVis == linkVis)
         {
-          // TODO: Open link right menu
+          g_modelRightMenu->Run(selectVis->GetName(), QCursor::pos(),
+              ModelRightMenu::EntityTypes::LINK);
         }
       }
     }