+++ gazebo/gui/model/ModelCreator.hh
@@ -211,12 +211,13 @@ namespace gazebo
       /// \brief Set the select state of a link.
       /// \param[in] _name Name of the link.
       /// \param[in] _selected True to select the link.
-      public: void SetSelected(const std::string &_name, bool selected);
+      public: void SetSelected(const std::string &_name, const bool selected);
 
       /// \brief Set the select state of a link visual.
       /// \param[in] _linkVis Pointer to the link visual.
       /// \param[in] _selected True to select the link.
-      public: void SetSelected(rendering::VisualPtr _linkVis, bool selected);
+      public: void SetSelected(rendering::VisualPtr _linkVis,
+          const bool selected);
 
       /// \brief Get current save state.
       /// \return Current save state.
@@ -296,7 +297,8 @@ namespace gazebo
       /// \param[in] _name Name of link.
       /// \param[in] _selected True if the link is selected, false if
       /// deselected.
-      private: void OnSetSelectedLink(const std::string &_name, bool _selected);
+      private: void OnSetSelectedLink(const std::string &_name,
+          const bool _selected);
 
       /// \brief Create link with default properties from a visual. This
       /// function creates a link that will become the parent of the
@@ -357,7 +359,7 @@ namespace gazebo
       private: void OnEntityScaleChanged(const std::string &_name,
           const math::Vector3 &_scale);
 
-      /// \brief Deselect all currently selected joint visuals.
+      /// \brief Deselect all currently selected link visuals.
       private: void DeselectAll();
 
       /// \brief Set visibilty of a visual recursively while storing their